# Optimize findlib cache